Created half of the organizer views, rearranged the filesystem and edited the nav graph
