+++ app/src/main/java/com/example/nostack/ui/attendee/AttendeeEvent.java
@@ -1,14 +1,15 @@
-package com.example.nostack;
+package com.example.nostack.ui.attendee;
 
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
-import androidx.navigation.NavHost;
 import androidx.navigation.fragment.NavHostFragment;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.nostack.R;
 
 /**
  * A simple {@link Fragment} subclass.
