+++ app/src/main/java/com/example/nostack/ui/organizer/OrganizerEvent.java
@@ -1,21 +1,21 @@
-package com.example.nostack;
+package com.example.nostack.ui.organizer;
 
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
-import androidx.navigation.NavHost;
-import androidx.navigation.fragment.NavHostFragment;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.nostack.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link StartUp#newInstance} factory method to
+ * Use the {@link OrganizerEvent#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class StartUp extends Fragment {
+public class OrganizerEvent extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -26,7 +26,7 @@ public class StartUp extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public StartUp() {
+    public OrganizerEvent() {
         // Required empty public constructor
     }
 
@@ -36,11 +36,11 @@ public class StartUp extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment StartUp.
+     * @return A new instance of fragment OrganizerEvent.
      */
     // TODO: Rename and change types and number of parameters
-    public static StartUp newInstance(String param1, String param2) {
-        StartUp fragment = new StartUp();
+    public static OrganizerEvent newInstance(String param1, String param2) {
+        OrganizerEvent fragment = new OrganizerEvent();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -60,27 +60,7 @@ public class StartUp extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-
-        View view = inflater.inflate(R.layout.fragment_start_up, container, false);
-
-        view.findViewById(R.id.AttendeeSignInButton).setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                NavHostFragment.findNavController(StartUp.this)
-                        .navigate(R.id.action_startUp_to_attendeeHome3);
-            }
-        });
-
-        view.findViewById(R.id.OrganizerSignInUpButton).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                NavHostFragment.findNavController(StartUp.this)
-                        .navigate(R.id.action_startUp_to_organizerHome);
-            }
-        });
-
-
-
         // Inflate the layout for this fragment
-        return view;
+        return inflater.inflate(R.layout.fragment_organizer_event, container, false);
     }
 }
