+++ app/src/main/java/com/example/nostack/ui/organizer/OrganizerSignIn.java
@@ -1,4 +1,4 @@
-package com.example.nostack.ui;
+package com.example.nostack.ui.organizer;
 
 import android.os.Bundle;
 
@@ -12,10 +12,10 @@ import com.example.nostack.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link OrganizerHome#newInstance} factory method to
+ * Use the {@link OrganizerSignIn#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class OrganizerHome extends Fragment {
+public class OrganizerSignIn extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -26,7 +26,7 @@ public class OrganizerHome extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public OrganizerHome() {
+    public OrganizerSignIn() {
         // Required empty public constructor
     }
 
@@ -36,11 +36,11 @@ public class OrganizerHome extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment organizer_home.
+     * @return A new instance of fragment OrganizerSignIn.
      */
     // TODO: Rename and change types and number of parameters
-    public static OrganizerHome newInstance(String param1, String param2) {
-        OrganizerHome fragment = new OrganizerHome();
+    public static OrganizerSignIn newInstance(String param1, String param2) {
+        OrganizerSignIn fragment = new OrganizerSignIn();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -61,6 +61,6 @@ public class OrganizerHome extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_organizer_home, container, false);
+        return inflater.inflate(R.layout.fragment_organizer_sign_in, container, false);
     }
 }
