+++ app/src/main/res/layout/fragment_organizer_home.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.OrganizerHome">
+    tools:context=".ui.organizer.OrganizerHome">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
@@ -15,8 +15,8 @@
             android:id="@+id/gridLayout1"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:rowCount="1"
             android:columnCount="2"
+            android:rowCount="1"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.10"
@@ -26,24 +26,24 @@
 
             <TextView
                 android:id="@+id/text_welcome"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
                 android:layout_row="0"
                 android:layout_column="0"
-                android:text="Welcome, "
-                android:textSize="17sp"
+                android:text="Welcome back, "
                 android:textColor="@color/white"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content" />
+                android:textSize="17sp" />
 
             <TextView
                 android:id="@+id/text_userWelcome"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
                 android:layout_row="0"
                 android:layout_column="1"
-                android:text="User"
-                android:textSize="17sp"
-                android:textColor="@color/yellow"
                 android:layout_marginStart="5dp"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content" />
+                android:text="User"
+                android:textColor="@color/yellow"
+                android:textSize="17sp" />
 
         </GridLayout>
 
@@ -51,7 +51,6 @@
             android:id="@+id/attendee_profileButton"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            app:srcCompat="@drawable/baseline_account_circle_24"
             android:backgroundTint="@color/transparent"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
@@ -59,6 +58,67 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintVertical_bias="0.05"
-            />
+            app:srcCompat="@drawable/baseline_account_circle_24"
+            android:contentDescription="Profile Button" />
+
+        <GridLayout
+            android:id="@+id/gridLayout2"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="25dp"
+            android:columnCount="2"
+            android:rowCount="1"
+            app:layout_constrainedWidth="false"
+            app:layout_constraintStart_toStartOf="@+id/gridLayout1"
+            app:layout_constraintTop_toBottomOf="@+id/gridLayout1">
+
+            <TextView
+                android:id="@+id/text_your"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_row="0"
+                android:layout_column="0"
+                android:text="Manage your"
+                android:textColor="@color/white"
+                android:textSize="17sp" />
+
+            <TextView
+                android:id="@+id/text_events"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_row="0"
+                android:layout_column="1"
+                android:layout_marginStart="5dp"
+                android:text="Events"
+                android:textColor="@color/lightBlue"
+                android:textSize="17sp" />
+
+        </GridLayout>
+
+        <ListView
+            android:layout_width="wrap_content"
+            android:layout_height="0dp"
+            android:layout_marginTop="8dp"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/gridLayout2" >
+        </ListView>
+
+        <Button
+            android:id="@+id/AddEventButton"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:backgroundTint="@color/lightBlue"
+            android:paddingHorizontal="55dp"
+            android:paddingVertical="12dp"
+            android:text="Create a New Event!"
+            android:textColor="@color/black"
+            android:textSize="20sp"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.50"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintVertical_bias="0.98" />
     </androidx.constraintlayout.widget.ConstraintLayout>
 </FrameLayout>
